+++ src/constants.js
@@ -1,1 +1,3 @@
+module.exports.typeError =
+  "Shescape requires strings or values that can be converted into a string using .toString()";
 module.exports.win32 = "win32";